requirements.txt
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
pickleshare==0.7.5
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
Werkzeug==1.0.1
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
tensorboard==2.3.0
tensorboard-plugin-wit>=1.6.0.post3
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
